MAGECLOUD-1031: Refactor InstallUpdate process
