+++ src/Magento/MagentoCloud/Process/Deploy/InstallUpdate/Update.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+namespace Magento\MagentoCloud\Process\Deploy\InstallUpdate;
+
+use Magento\MagentoCloud\Process\ProcessInterface;
+
+class Update implements ProcessInterface
+{
+    /**
+     * @var ProcessInterface
+     */
+    private $process;
+
+    /**
+     * @param ProcessInterface $process
+     */
+    public function __construct(
+        ProcessInterface $process
+    ) {
+        $this->process = $process;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function execute()
+    {
+        $this->process->execute();
+    }
+}
